requirements.txt
# Core LLM & Google Gemini
google-genai==1.50.1

# HTTP requests
requests>=2.32.0,<3.0.0
httpx>=0.28.1,<1.0.0

# Authentication
google-auth>=2.14.1,<3.0.0

# Data validation & typing
pydantic>=2.9.0,<3.0.0
typing-extensions>=4.11.0,<5.0.0
annotated-types>=0.6.0

# Async handling
anyio>=4.8.0,<5.0.0
websockets>=13.0.0,<15.1.0

# Retry mechanism
tenacity>=8.2.3,<9.2.0
